refactor(nixl): extract pega rdma v1 helpers
